Add BirthDate, Note, Participant java files
Participant class is the class to represent person coming to the event.
Participant will have extra fields including birthDate, notes, and nextOfKins specifically catered for contact purpose and
refined to suit event organizer needs.

BirthDate have getAge() method which return age of type Int to check if a participant is of valid age for certain event or not
Note have different level of Importance in Note.Importance enum

Things to note:
**Participant extends from Person class
**Participant currently have the same constructor as Person class and use withXXX builder chaining method
   as setter in to return new Participant
**Include Test for 3 different classes