Add Note Command
Right now, we cannot take notes on people in the Address Book, which some users may want to do.

To fix this, we add a new class Note to add a note field to the person model, as well as NoteCommand and NoteCommandParser to handle the new note command. Finally, we edit our other classes to support the operations required to process, track and save notes.

This allows users to put down a simple note for each of the people in the Address Book, to further identify them.