Update scalability/capacity NFR
Assume a tutorial/lab class of maximum about 20 students like in CS2100, and that a TA could be teaching multiple lab sessions + another module. 
We then take 100 as the upper limit of the number of students that we expect a teaching assistant (The user) to have.
If we stress test in the future, we should take this as the maximum number of students to have.