Update Person class duplicate checks
Person class isSamePerson method currently returns true if
the two Person objects have the same name.

The Receptionist may encounter people with the same name,
and this can be an issue as they cannot add a person with
the same name.

Updating the isSamePerson method to adopt a less strict
check for duplicates by comparing not just the name but
the phone and email field as well. This means, this
method will only return true when the two Person objects
have the same name, email and phone field.

A person is unlikely to have the same name, phone and email
as another person. If the Receptionist try to add such
a person while such a person already exists, the prompt
will be helpful to identify if this person is already
registered in the system.