Add Patient functionality to program
Person class is currently used as the main object
in the entire program.

As this project is being adapted to fit PlannerMD,
the main object of concern should be Patient,
which is an extension of the Person class with an
additional Risk field.

Some considerations took in place for this refactoring
was the possible addition of other classes like Doctor,
thus files like AddCommand.java are renamed to
AddPatientCommand.java and AddCommand.java becomes
an abstract class containing the keyword that
is used to add both types of classes. This is done
in consideration for adding mass operations
such as mass deleting different classes such as
Patients and Doctors.

Some noteworthy changes include changes to the JSON format
for Storage. Note that the key changed from "persons" to
"patients" in JsonSerializablePlannerMd.java. A "risk" key
is also added to JsonAdaptedPatient.java (previously
JsonAdaptedPerson.java).

The class NameContainsKeywordsPredicate
has also become a generic class in which the class
becomes NameContainsKeywordsPredicate<T extends Person>
implements Predicate<T>. This is to support the same
predicate checks for other classes with the name fields,
i.e. other classes which inherit from the Person class.

Note that at this commit the program is able to run
and perform the required funtions. The test suites
will be fixed in the subsequent commits to fix
the failing tests.