Add Search Feature
Search Command: Generic and more functional find Command

A more functional find Command for the user to search with loose matching and
also by client's attribute as while.

Let's add SearchCommand along with the resepective parser and a predicate used for
updating the model.

User guide and Test case has been updated for this feature.