Add retrieval of group in model for Add command
When adding a student, the AddCommand object is now created with a dummy Group in the Student object.

This is done so that the AddCommand#execute stage will be the one to do the checking if the indicated group (with respective groupName) exists in the model or not. It will then search for within the model, and replace the group in the parsed student to the correct group object stored in the model. This is because this is the stage that should have visibility of model, not at the parsing stage.

Note that this behaviour relies on Group#isSameGroup only checking for whether groupName is equal.