Refactor COMMAND_WORD into a enum
One Command may have many different alias. Currently, to add more
alias to a command, we have to add more COMMAND_WORD in each
command class. This will lead to more switch case in AddressBookParser
and this is not scalable in the future if there were to be many more
aliases added.

Refactoring all the COMMAND_WORD in the different command into one
enum CommandWord allows us consolidates all the COMMAND_WORD for
different Command. The use of enum also removes the need for adding
more switch case everytime a new alias is added. We just have to add
the alias in the enum and the switch block in AddressBookParser will
still work.