Update UI to display Persons properly
There are new properties of Person. UI needs to display the new
properties properly.

UI also needs to display a symbol for gender. It is appropriate to let
the enum GenderType handle its symbols.

Let's update the FXML and code handling the display of Person. Let's
also move enum GenderType to a new file on its own. Doing so allows it
to be public and has public fields that are accessible by PersonCard.