Add functionality to view list of instructions.
The current help command only shows the link to the user guide and
makes it necessary for a new user to go outside of the application to
know how to start using it.

This makes getting to know the basics of the application unnecessarily
tedious as the user has to go to an external source of information.

Implementing a list of instructions inside the app will make the
process of trying to understand the basics as a new user much simpler.