Update Untag to throw exception with invalid tags
Untag does not throw error if tags that does not exist or empty tags are
entered in the command.

Command should not be allowed to execute if non-existent tags or empty
tags are entered.

Let's allow Untag to detect non-existent tags and empty tags and throw
errors if detected.

Throwing errors in this situation is preferble as it does not make sense
for the command to execute.