Add RemarkClassCommandParser
Currently, the user is required to type "remarkclass" to write a
note for a particular class. This change needs to be reflected in the
user guide.

Future developments will work on allowing users to input an optional
remark when creating a class.

EditClassCommand should not handle the update of class remarks as that
will overlap significantly with the RemarkClassCommand's responsibility.

Minor improvement in GUI for Tuition Card: specify "capacity" and "notes"
so users know what each detail means.