Rate Command: Improve GUI rendering
Current rating is displayed plainly as an integer under each PersonCard. Does not stand out to user and can confuse user without any background knowledge of the fields.

A highlighted box with bolded text and a star emoticon will help users easily and intuitively understand what the field stands for, and aids our rating field in achieving what it is supposed to: readability and comprehensibility.

Let's
* Improve the look of the rating field in the GUI