Fix for bug caused by delete command
When a Participant is deleted from Managera using the delete command, it
is removed from the Participant list but not from any events that the
Participant was previously attending.

Hence, when Managera is restarted, it still attempts to restore the now
missing Participant object in the event(s). This causes Managera to not
start up at all.

Let's add references to any events in the Participant class so that when
a Participant object is deleted, we can remove any references to this
object in the events the Participant was attending. These references are
restored on startup when the Event objects are restored from the save
file. The event references are not stored in the save file.

* Add a list of events as an attribute to Participant class.
* Fix the bug causing Managera to not start after a Participant of an
  event is deleted.
* Changes to certain methods which violate Law of Demeter.
* Add test cases for new methods to add and remove event references from
  Participant objects.
* Update user guide as some command syntax are outdated.