Refactor: Refactor Command Parsing and Execution
Right now, we use a very simple command system, where each command has its own Parser and Command class, and there is no inheritance or OOP within these. This is not sufficient as we want to implement subcommands to our applications based on the Objects we will create.

To support this functionality, we add an Executor with the current Model instance to allow the execution of commands to be separated from the classes. We also create the TargetedCommand, GroupCommand and PersonCommand classes that are more specific versions of the Command class.

This allows us to specify the target of the command after parsing the first command word. We can then do the remaining parsing and execution afterwards.

3 classes need to be added for each new person Command: a class that extends `PersonCommandParser`, a class that extends `PersonExecutor` and a class that extends `PersonCommand`.

Fix javadocs.
Fix bugs.

TODO: Make sure that all person/group specific commands follow the correct format.