Resolve issue with double Participant objects
Upon startup, Managera would create up to two instances of each
Participant. One instance is added into addressbook, and the other is
added into the Participant list of any event the Participant is
attending. This is due to an outdated method of deserialising the
Participant and Event objects from the save file.

This causes unneccessary usage of extra memory to store the extra
Participant objects.

Let's adjust the way that Managera deserialises the Participant and
Event objects from the save file so that only one Participant object
is ever created for each Participant.