Fix UI bug for events not updating participants
When Participants are added to or removed from an Event, the Event in
the Event list of Managera does not update its list of Participants
correctly. This is due to how ObservableList works, which listens to
changes in the objects in the list, but not when the attributes of the
objects are changed.

Let's change the list of Participants in Events from an ArrayList to an
ObservableList, and include a callback in UniqueEventList to listen for
any changes made to this ObservableList in the Event. That way, when a
Participant is removed from or added to an Event, it would invoke the
callback method and cause javaFX to update the UI accordingly.