Add more unit tests for command classes
Add the following test files:
* RemoveEventCommandTest
* RemoveParticipantFromEventCommandTest

Modify the following test files:
* AddParticipantToEventCommandTest

Test coverage for all classes in the logic.commands package is 100% as
reflected in IntelliJ, with the exception of EditCommand. (Due to the
incomplete implementation of EditCommandDescriptor)

Let's add an EventBuilder class to help create new Event objects
for testing. This will be useful for testing commands that would modify
the attributes of the Event object.

Let's also add an EventNameContainsKeywordPredicate class. This class
will be useful later when implementing the feature to allow users to
serach for Events by name.