Create Appointment class
Create AppointmentDateTime class to encapsulate the starting/ending
date and time of an appointment

The Appointment class has both starting and ending date/time so that it
can be checked against other appointments involving the same patient or
doctor to detect clashes. New appointments that clash with existing ones
should not be allowed to be added.