Refactor 'add' in UniqueFriendsList and UniqueGamesList
'add' method in UniqueFriendsList and UniqueGamesList
uses 'contains' to check if the object is already in the list.
'contains' uses 'equals' to check, wherein the 'equals' checks if all
attributes are equal. This allows for insertion of multiple objects with same
ids.

Fix this by modifying add to use 'containsId' instead to check for duplicates.