Set member availability command: Improve parsing and ease of use
Currently, availability values are implemented as a string.
This brings several problems. The string is hard to sort, hard to
parse and verify validity. Having to type in the names of days is
also tedious when the user could just type "d/1 2 3" instead to
represent the days. Using numbers which can be converted to
DayOfWeek makes it easier to parse and validate.

Let's use java.time.DayOfWeek and change the data type of
availability values from String to List<DayOfWeek> as well as change
the command format to improve ease of use and validation.