Add Trie Functionality
Currently, our Trie implementation only supports finding the first item or the first item based on a query. This is insufficient if we are to use it for our Trie command.

Therefore, we add the findAllMatches command to return us a list of all the matches in our Trie.

This allows us to easily display whole lists filtered by a search string.