Add feature to view a Participant's details
Let's add a new feature to Managera to allow users to search for
a Participant using their participantId and view their details. If the
user's input is not precise enough and there are multiple participantIds
that match, these IDs will be displayed to the user so that he may
refine his search.

Let's add:
* A view command to allow users to find a Participant by their
  participantId and view the Participant's details.
* Update user guide to reflect this new feature.
* Fix issues with the user guide regarding broken hyperlinks and
  outdated syntax.
* Add test files for the newly added classes: ViewCommand,
  ViewCommandParser, ParticipantIdMatchesGivenIdPredicate. (100% line
  coverage)
* Add an additional test for AddCommandTest to provide coverage for
  toString method.