Change implementation of view command
I realised that with my previous implementation, the user would
encounter an issue if there are at least two participants where one
participant's ID is contained within the other. The user would never be
able to narrow down his search to the former. The implementation has
been adjusted to fix that issue, but the command will no longer list the
similar IDs as with the previous implementation.