Refactor CommandResult into its different types
Currently, CommandResult is getting too cluttered as more functionality is being added. For instance, with the addition of the Download command, there would be 3 different constructors and an excessive number of attributes for the CommandResult class.

Hence, different classes were created for each CommandResult type, to better encapsulate the different cases where we need to change the UI on some command provided by the user. Overall, this should make the code less verbose and easier to understand and maintain.