Sort contacts by alphabetical order
This commit creates a SortedList that wraps around the existing
FilteredList of ModelManager, which updates automatically as
FilteredList updates.

As a consequence of this change, getFilteredPersonList() in Model has
been renamed to getSortedPersonList(). An additional method,
updateSortedPersonList(), has been created for future developers to set
specific comparators on the SortedList. Also, a natural order has been
created for the Person and Name classes, based on case-insensitive
lexicographic order of full names.