UI Logic: update weekly panel implementation
The previous logic implementation uses global static variables that
increases coupling and is hard to trace.

Creating a WeeklyTasks class to handle filtering of task reduces
coupling and improves overall code quality.

It is now easier to manage the updating of the weekly panel through
model manager and show command.