Fix EditCommandBug
A bug exists when using EditCommand such that if a Participant is edited
while linked to Events, the Participant is not updated in those Events,
hence the Participant subsequently cannot be removed from the Event and
restarting Managera will result in an inability to load from the save
file.

Let's edit Event, Participant and ModelManager such that editing a
Participant will cause the links it has to Events to be transferred to
the new edited Participant.