Update view and done commands
View command uses participantId to find the specific participant, but
this adds little value as it only searches within the filtered list. It
would be enough (and easier for the user) to just use the displayed
index to view the Participant.

The view command currently is also not very useful as the ParticipantCard
already displays most of the details of the Participant.

The doneEvent command can be jsut simplified to just `done` as it is an
Event-specific command.

Let's:
* Change view command to use index to perform the search instead.
* Give view command the ability to show the events that the Participant
  is attending. This will give value to the view command.
* Adjust the test files to accomodate the changes above.
* Change the command word of the done command.
* Change the implementation of done command so that it aligns with
  existing implementations better.