Add allocate member command
Add command to allocate member to a facility on a given day.
Checks for valid member and facility indices given, whether
the member is available on that day, whether the facility
is at max capacity and whether the person is already allocated to
the facility. However, does not check whether member is
assigned to different facilities on the same day and each member
can only be allocated to a facility once regardless of day.