Implement directional sorting in Sort command
This commit modifies the previous approach of input parsing to use
ArgumentMultimap. This required a change to the internal data structure
from HashMap to LinkedHashMap, which would preserve the insertion order
of prefixes. An additional method, getPrefixes(), has been created in
ArgumentMultimap to return a list of prefixes ordered by index of last
occurrence.