Update delete feature to allow multiple deletions at one go
Currently, the delete command only accepts a string containing one task number for deletion. It does not support the deletion of multiple tasks at once.

In order to make things more efficient, multiple tasks can now be deleted at once such that users will not have to enter the same command a few times over should they decide to repeat the delete command.

When multiple indexes are entered by user, the DeleteCommandParser class parses the user's input (String) into a List<Index> of multiple Index objects, containing the index of all the tasks to be deleted. As per before, the DeleteCommandParser creates a DeleteCommand object, but this time taking in the List<Index> instead.

The DeleteCommand class is changed such that its constructor now accepts a List<Index> instead of just an Index to facilitate this deletion of multiple tasks at once. In its execute method, it deletes the items iteratively from the last index to prevent future deletions from operating on wrong indexes due to the task list updating itself in each iteration.