Bug Fixed
Bug Description:

When the User Name was being updated
via the edit command, the name on the
menu bar did not change, however
internally the name was changed. On
restarting the application, the changed/new
name could be observed.

Bug Fixed:

Now, as the User updates their name,
the changes on the menu bar are also
reflected immediately.