Bugfix mainApp
It is possible to start with a sample schedule, and the user's own contacts.
This would show discrepancies between the schedule and the user's contacts.

Using a boolean to track whether sampleSchedule was used, allows us to ensure
that if there is a case where schedule.json was missing, but contacts.json was
found, we would not have discrepancies between the two lists.