Allow Duplicate Names
Previously, entries with the same Name is treated as the same person.

According to @kwokieee, this is a bug as it prevents me from adding friends whose full name I may not know (e.g multiple 'John's).

To solve this, I have changed the method which check if 2 persons are the same to follow the 'equals' method. This means 2 persons are duplicate, only if all details of them are the same, which sounds like a valid reason to give.