Fix Index's Integer.MAX_INT Bug
All commands which involves the parsing of Index will return an Invalid Command Format when a Non-Zero Integer cannot be parsed.

According to @xianlinc, the correct error should be 'The person index provided is invalid', no matter how large the input number. [Valid point]

I've added a check, such that all string that
1. Contains just numbers AND
2. Not the integer range (i.e between -2 billion to 2 billion) AND
3. Is a positive number (any length), but must not be all 0s...
will return Integer.MAX_INT.
This assumes that the users will not have 2.147 billion friends in Socius, which should be a reasonable assumption to make (Consider the size of json file and time taken to import!!)