Improve consistency of error messages
Adapted isNonZeroUnsignedInteger method in StringUtil to isNumeric.
In event of a numeric string which cannot be parsed to an int, the
exception is caught separately from non-numeric strings.
Hence, when the index out of bounds, negative, or fractional is
provided, the error message is "invalid index" rather than
"invalid command format".