Lehao - Undo command bug fixes (#219)
* Lehao - Undo command bug fixes

The previous undo has such a bug: when delete person and then undo, the client list of the affected appointment will not be refreshed (undo). This is because a state simply copy the reference of the schedule, and different schedules in different states actually refer to the same appointment and same client list. To solve this we need to create deep copy when initialize the state and get schedule of the state.

* Lehao - undo command bug fixes

* Lehao - undo command bug fixes

* Lehao - undo command bug fixes

undo command success message now contains the command name.
undo command now restore the predicates correctly.

* Lehao - Edit the find command undo message

Co-authored-by: Li-Lehao <e0492462@u.nus.edu>